==> scripts/0_age.ipynb <==
# Load Data


import pandas as pd

# df_baseline = pd.read_csv(r'../data/FBaseline.txt', sep='|')                         # no
# df_adv_events = pd.read_csv(r'../data/FAdvEvent.txt', sep='|')                       # no
df_roster = pd.read_csv(r'../data/FPtRoster.txt', sep='|')                           # no
# df_roster= df_roster[df_roster['FPtStatus'] != 'Dropped']
# df_lab_data = pd.read_csv(r'../data/FLabData.txt', sep='|')                          # no
df_medical_conditions = pd.read_csv(r'../data/FMedicalConditions.txt', sep='|')
df_medications = pd.read_csv(r'../data/FMedications.txt', sep='|')

# df_bgm = pd.read_csv(r'../data/FDataBGM.txt', sep='|')                              # no
# df_cgm = pd.read_csv(r'../data/FDataCGM.txt', sep='|')                              # no
# df_acc = pd.read_csv(r'../data/FDataAccel.txt', sep='|')                            # no
# df_acc_info = pd.read_csv(r'../data/FDataAccelSubjectInfo.txt', sep='|')            # yes for 148 and not 205
# df_daily_log = pd.read_csv(r'../data/FDailyLog.txt', sep='|')                       # no
# df_insulin = pd.read_csv(r'../data/FInsulin.txt', sep='|')                          # no

df_visit_final = pd.read_csv(r'../data/FFinalVisit.txt', sep=',')
# df_visit_follow_up = pd.read_csv(r'../data/FFollowUpVisit.txt', sep=',')            # no
# df_visit_final_screening = pd.read_csv(r'../data/FScreeningFinalVisit.txt', sep='|')# no
# df_visit_info = pd.read_csv(r'../data/FVisitInfo.txt', sep='|')                     # no


# df_final_status = pd.read_csv(r'../data/FPtFinalStatus.txt', sep='|')               # no
df_sample_results = pd.read_csv(r'../data/FSampleResults.txt', sep='|')





# Get All names of All Volumns for All DataFrames

# columns_baseline = df_baseline.columns.tolist()
# columns_adv_events = df_adv_events.columns.tolist()
# columns_roster = df_roster.columns.tolist()
# columns_lab_data = df_lab_data.columns.tolist()
columns_medications = df_medications.columns.tolist()
columns_medical_conditions = df_medical_conditions.columns.tolist()

# columns_bgm = df_bgm.columns.tolist()
# columns_cgm = df_cgm.columns.tolist()
# columns_acc = df_acc.columns.tolist()
# columns_acc_info = df_acc_info.columns.tolist()
# columns_insulin = df_insulin.columns.tolist()
# columns_daily_log = df_daily_log.columns.tolist()


# columns_visit_final_screening = df_visit_final_screening.columns.tolist()
columns_visit_final = df_visit_final.columns.tolist()
# columns_visit_follow_up = df_visit_follow_up.columns.tolist()
# columns_visit_info = df_visit_info.columns.tolist()

columns_sample_results = df_sample_results.columns.tolist()
# columns_final_status = df_final_status.columns.tolist()




# Removes all the dropped patients



df_roster= df_roster[df_roster['FPtStatus'] != 'Dropped']
dataframe_names_to_filter = [
    'df_baseline', 'df_adv_events', 'df_lab_data', 
    'df_medications', 'df_medical_conditions', 'df_bgm', 'df_cgm', 
    'df_acc', 'df_acc_info', 'df_insulin', 'df_daily_log', 
    'df_visit_final_screening', 'df_visit_final', 'df_visit_follow_up', 
    'df_visit_info', 'df_sample_results', 'df_final_status'
]

for name in dataframe_names_to_filter:
    if 'PtID' in globals()[name].columns:
        globals()[name] = globals()[name][globals()[name]['PtID'].isin(df_roster['PtID'])]
        

# List of DataFrame variable names as strings
dataframe_names = [
    'df_baseline', 'df_adv_events', 'df_roster', 'df_lab_data', 
    'df_medications', 'df_medical_conditions', 'df_bgm', 'df_cgm', 
    'df_acc', 'df_acc_info', 'df_insulin', 'df_daily_log', 
    'df_visit_final_screening', 'df_visit_final', 'df_visit_follow_up', 
    'df_visit_info', 'df_sample_results', 'df_final_status'
]

# Loop through each DataFrame name, get its columns, and assign to a new variable
for name in dataframe_names:
    columns_name = f"columns_{name[3:]}"  # Create the variable name for columns
    globals()[columns_name] = globals()[name].columns.tolist()
    
# Now, variables like columns_baseline, columns_adv_events, etc., are defined
# Loop through each DataFrame name, extract unique 'PtID', and assign to a new variable
for name2 in dataframe_names:
    if 'PtID' in globals()[name2].columns:  # Check if 'PtID' exists in DataFrame
        unique_name = f"unique_{name2}"  # Create the variable name for unique 'PtID' values
        globals()[unique_name] = globals()[name2]['PtID'].unique()
