==> scripts/0_jobscript_dict_remove_ptid.ipynb <==
import os

# Path to the directory
# folder_path = "/Users/au605715/Documents/GitHub/jchr_racial_diff/Data/2_1_1_predicted_results_rnn_wb/2_1_1_predicted_results_rnn_wb_v3"
folder_path = '/home/hbt/jchr_data/jchr_racial_diff/results/processed_data/2_1_1_predicted_results_rnn_wb/2_1_1_predicted_results_rnn_wb_v4'
# folder_path = '/home/hbt/jchr_data/jchr_racial_diff/results/processed_data/2_1_2_predicted_results_rnn_mf'
# List all entries in the directory
all_entries = os.listdir(folder_path)

# Filter out directories, keep only files
files = [f for f in all_entries if os.path.isfile(os.path.join(folder_path, f)) and f.endswith('.pkl') and f.startswith('patient')]

import pickle
# file_path = "/Users/au605715/Documents/GitHub/jchr_racial_diff/Data/processed_data/1_3_1_data_split_wb_v3.pkl"
file_path = "/home/hbt/jchr_data/jchr_racial_diff/results/preprocessed_data/1_3_1_data_split_wb_v3.pkl"
# file_path = '/home/hbt/jchr_data/jchr_racial_diff/results/preprocessed_data/1_3_2_data_split_mf_v3.pkl'
# Read from file
with open(file_path, 'rb') as file:
    dictionary = pickle.load(file)
    
    
my_dict = dictionary

# Function to extract PtID and ratio from the filename
def extract_info(filename):
    base_name = os.path.splitext(filename)[0]  # Remove the extension
    parts = base_name.split('_')
    PtID = parts[0].replace('patient', '')  # Extract PtID
    ratio = parts[1].replace('ratio', '')  # Extract ratio
    return int(PtID), int(ratio)

# Iterate over the file list and remove corresponding entries from the dictionary
for i_file in files:
    key = extract_info(i_file)
    if key in my_dict:
        del my_dict[key]


# file_path = "/home/hbt/jchr_data/jchr_racial_diff/results/processed_data/2_1_2_predicted_results_rnn_mf/run2.pkl"


# with open(file_path, 'wb') as file:
#     # Serialize and save the list to the file
#     pickle.dump(my_dict, file)
