==> scripts/1_2_preprocess_feature_extraction.ipynb <==
"""_summary_
Creates observational windows with target
"""

#%% import data
import pandas as pd
import my_utils
df = pd.read_csv(r'/home/hbt/jchr_data/jchr_racial_diff/results/preprocessed_data/1_1_racial_diff_dataclean.csv')


df_roster = pd.read_csv(r'/home/hbt/jchr_data/_data/FPtRoster.txt', sep='|')
df_roster.drop(columns=['RecID','SiteID'], inplace = True)
df_roster['RaceProtF'] = df_roster['RaceProtF'].replace('Non-Hispanic Black', 'black')
df_roster['RaceProtF'] = df_roster['RaceProtF'].replace('Non-Hispanic White', 'white')
df_roster.rename(columns={'RaceProtF':"Race"}, inplace=True)

df_baseline =pd.read_csv(r'/home/hbt/jchr_data/_data/FBaseline.txt', sep='|')


# create sample of 1 hour and prediction horizon of 1 hourho
df_feature= my_utils.feature_extraction_dl(df,window_size=4, prediction_horizon=7, col_patient_id='PtID', col_glucose='CGM')


#%% add ethnicity
df_race = pd.DataFrame(data=[df_roster['PtID'],df_roster['Race']]).transpose()
df_feature = pd.merge(df_feature, df_race, on='PtID', how='left')

#%% add educational level

df_baseline = pd.DataFrame(data=[df_baseline['PtID'],df_baseline['Gender'],df_baseline['EduLevel']]).transpose()
df_feature= pd.merge(df_feature, df_baseline, on='PtID', how='left')


#%% For saving dataframes
numb_ptid=df_feature.PtID.unique()
print(len(numb_ptid))

df_feature.head()



df_feature.to_csv(r'../results/preprocessed_data/1_2_model_input_ws60min_ph60min.csv', index=False) # takes forever, maybe try with chunk_size = 10000 next time
print('finished')



df_feature.PtID.unique()
