==> scripts/3_1_calc_results_naive_approach_v2.ipynb <==
import pandas as pd
import numpy as np
from decimal import Decimal, ROUND_HALF_UP

# Load the DataFrame
df = pd.read_csv(r'../results/processed_data/2_lmem_results/3_1_calculated_results_v6_race_lmem_v2.csv')
df.drop(columns=['Unnamed: 0', 'ptid_test'], axis=1, inplace=True)

# Filter the DataFrame for ratio = 10 and then by Race and Gender
# df_ratio_10_black = df[(df['ratio'] == 10) & (df['Race'] == 'black') & (df['Gender'] == "M")& (df['AgeGroup']==1)]
# df_ratio_10_white = df[(df['ratio'] == 10) & (df['Race'] == 'white') & (df['Gender'] == "M")& (df['AgeGroup']==1)]

df_ratio_10_black = df[(df['ratio'] == 10) & (df['Race'] == 'black')]
df_ratio_10_white = df[(df['ratio'] == 10) & (df['Race'] == 'white')]








# Define function to calculate mean and 95% CI
def calculate_mean_and_95ci(df, column):
    # Convert the column to Decimal for precise rounding
    decimal_values = df[column].apply(lambda x: Decimal(str(x)))
    
    mean = sum(decimal_values) / Decimal(len(decimal_values))
    std = (sum((x - mean) ** 2 for x in decimal_values) / Decimal(len(decimal_values) - 1)).sqrt()
    n = Decimal(len(decimal_values))
    
    standard_error = std / n.sqrt()
    ci_95 = Decimal('1.96') * standard_error
    
    lower_bound = mean - ci_95
    upper_bound = mean + ci_95
    
    # Round the values
    rounded_mean = mean.quantize(Decimal('0.01'), rounding=ROUND_HALF_UP)
    rounded_lower = lower_bound.quantize(Decimal('0.01'), rounding=ROUND_HALF_UP)
    rounded_upper = upper_bound.quantize(Decimal('0.01'), rounding=ROUND_HALF_UP)
    
    print(f"{rounded_mean} [ {rounded_lower} - {rounded_upper} ]")
    
    return {
        'Mean': rounded_mean,
        '95% CI Lower Bound': rounded_lower,
        '95% CI Upper Bound': rounded_upper
    }

print('white')
# Results for White race
results_white = calculate_mean_and_95ci(df_ratio_10_white, 'rmse_naive')

print('black')
# Results for Black race
results_black = calculate_mean_and_95ci(df_ratio_10_black, 'rmse_naive')








import math

# From describe() output
mean_white = df_ratio_10_white['rmse_naive'].mean()
std_white = df_ratio_10_white['rmse_naive'].std()
n_white = len(df_ratio_10_white)

se_white = std_white / math.sqrt(n_white)
ci_95_white = 1.96 * se_white

lower_bound_white = mean_white - ci_95_white
upper_bound_white = mean_white + ci_95_white

print(f"White CI from describe(): {mean_white} [ {lower_bound_white} - {upper_bound_white} ]")



import math

# From describe() output
mean_black = df_ratio_10_black['rmse_single'].mean()
std_black = df_ratio_10_black['rmse_single'].std()
n_black = len(df_ratio_10_black)

se_black = std_black / math.sqrt(n_black)
ci_95_black = 1.96 * se_white

lower_bound_black = mean_black - ci_95_black
upper_bound_black = mean_black + ci_95_black

print(f"Black CI from describe(): {mean_black} [ {lower_bound_black} - {upper_bound_black} ]")
